chore: split tests from source
