+++ runtime/parachains/src/paras.rs
@@ -24,7 +24,7 @@
 //! only occur at session boundaries.
 
 use sp_std::prelude::*;
-// use sp_std::marker::PhantomData;
+use sp_std::marker::PhantomData;
 use sp_runtime::traits::{One, BlakeTwo256, Hash as HashT, Saturating};
 use primitives::v1::{
 	Id as ParaId, ValidationCode, HeadData, LocalValidationData,
